Finished task 2nd part

+++ my-app/src/components/use-request-delite-task.js
@@ -1,22 +1,27 @@
 import { useState } from 'react';
 
-export const useRequestDeleteTask = (refreshTasks, setRefreshTasks) => {
+export const useRequestDeleteTask = (refreshTasks, setRefreshTasks, id, setShowInput) => {
 	const [isDeleting, setIsDeleting] = useState(false);
 
 	const requestDeleteTask = (item) => {
 		setIsDeleting(true);
-		fetch(`http://localhost:3030/task/${item.target.id}`, {
-			method: 'DELETE',
-		})
-			.then((rawResponse) => rawResponse.json())
-			.then((response) => {
-				console.log('Удален, ответ от сервера: ', response);
-				setRefreshTasks(!refreshTasks);
-			})
-			.finally(() => {
-				setIsDeleting(false);
-			});
-	};
+		if (id !== 'add') {
+			console.log(id);
 
+			fetch(`http://localhost:3030/task/${item.target.id.slice(4)}`, {
+				method: 'DELETE',
+			})
+				.then((rawResponse) => rawResponse.json())
+				.then((response) => {
+					console.log('Удален, ответ от сервера: ', response);
+					setRefreshTasks(!refreshTasks);
+				})
+				.finally(() => {
+					setIsDeleting(false);
+				});
+		} else {
+			setShowInput(false);
+		}
+	};
 	return { requestDeleteTask, isDeleting };
 };
